#6 FAQ

+++ src/main/java/com/universityweb/course/repository/ReviewRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-    List<Review> findByCourseId(int courseId);
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+    List<Review> findByCourseId(Long courseId);
 }
